requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
PyMySQL
flask-bcrypt
PyJWT
